[fix:test] Force postAceInit hook to be bundled
The `postAceInit` hook was created here to provide a way to access the
bundled jQuery instance on the tests of other plugins, so those tests
can simulate jQuery events.

Because of that reason `postAceInit` **must** be bundled, while the
other hooks of this plugin should not, so we keep them on a separated
part.