[perf] Split CSS and JS entry files to speed up JS bundling
Create separated chunks for JS and CSS, so we don't need to rebuild CSS
files if we're changing only the JS part. As changes on JS is much more
frequent -- and need more experimenting on the browser --, this is the
most important part to improve performance.

Improvements (for a bundle of 42 CSS files and 50 JS files, changing a
single CSS or JS file): from 8s to less than 2s.