[fix] Split webpack.config.js files depending on the Etherpad settings
Create different `webpack.config.*.js` files depending of the (current)
4 combinations of settings: with/without minimization; and with/without
CSS bundling. This avoids providing webpack with options it won't need,
like optimizing CSS when we're not minimizing them.

Also revert part of previous commit, and do not split CSS and JS files
into different entries. This change might decrease the performance of a
config with CSS bundling *and* minimizing, but it increases the
performance of the most common config for the dev environment (where
watch mode is usually running): CSS bundling *without* minimizing files.

Improvements (for a bundle of 42 CSS files and 50 JS files, changing a
single CSS or JS file): from 2s to 0.5s.