[feat] Run webpack only a file changes
On this commit we change the way the webpack runs. Before, every 1
second we had to check if a file has changed. The main reason of this
approach was because we were using virtualbox to run docker. Webpack has
some issues with this setup, please check:
https://webpack.js.org/configuration/watch/#watchoptionspoll
For users that use virtualbox, they need to enable the "pool" flag.